// website-next/src/lib/client/types.gen.ts
export type KillDistributionPointResponse = {
    bucketStart?: string;
    kills?: number;
};
// ...
export type ScoreboardEntryResponse = {
    playerName?: string;
    kills?: number;
};

export const ScoreboardPeriod = {
    DAILY: 'DAILY',
    WEEKLY: 'WEEKLY',
    MONTHLY: 'MONTHLY',
    ALL_TIME: 'ALL_TIME'
} as const;

export type ScoreboardPeriod = typeof ScoreboardPeriod[keyof typeof ScoreboardPeriod];